cart view and remove from cart implemented with database integration
loading cart view with a student ID and sections in that student's cart should read from the database. to test try logging in, adding courses to your cart, then hard-coding studentID in cartPanel to your studentID and the courses should appear when loading the screen from debug panel. removing courses from the cart should remove them from both the screen and the database as well